create collection classes, commence NewEventVC
